edit styles

+++ src/components/Secondary Pages/Cards/PackageCard.js
@@ -7,13 +7,14 @@ import { useEffect, useState } from "react";
 import Swal from "sweetalert2";
 import { useContext } from "react";
 import { MyContext } from "../../Main Pages/Redux";
-const PackageCard = (props) => {
+const PackageCard = ({cardData}) => {
   const personData = useContext(MyContext);
+  console.log("cardDa",cardData)
+  console.log("check",personData==="admin")
   const navigate = useNavigate();
-  const cardData = props.cardData;
   const [showHallDetails, setShowHallDetails] = useState(false);
   const [hall, setHall] = useState();
-
+const nav=useNavigate();
   const allCardData = () => {
     fetch(`http://127.0.0.1:8000/api/auth/getHall/${cardData.hall_id}`)
       .then((response) => response.json())
@@ -26,40 +27,52 @@ const PackageCard = (props) => {
   };
   function handleClick() {
     navigate(`/EditPackage/${hall.id}`, {
-      state: { hall: hall },
+      state: { hall: cardData },
     });
   }
 
-  function handleHallDetailsClick() {
-    setShowHallDetails(true);
+  function handleHallDetailsClick(hall) {
+    // setShowHallDetails(true);
+    nav(`/hallDetails/${cardData.hall_id}`);
   }
 
   function deletePackage() {
-    fetch(`http://127.0.0.1:8000/admin/auth/destroyOffer/${hall.id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${personData.token}`,
-        "auth-token": `${personData.token}`,
-      },
-    })
-      .then((response) => {
-        if (!response.ok) {
-          throw new Error("Network response was not ok");
-        }
-        return response.json();
-      })
-      .then((data) => {
-        Swal.fire({
-          title: "Package has been deleted",
-          showCancelButton: false,
-        });
-        window.location.reload();
-      })
-      .catch((error) => {
-        console.error("Error fetching data:", error);
-      });
+    Swal.fire({
+      title: "Are you sure you want to delete this package?",
+      showCancelButton: true,
+      confirmButtonText: "Delete",
+      cancelButtonText: "Cancel",
+      icon: "warning",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://127.0.0.1:8000/admin/auth/destroyOffer/${cardData.id}`, {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: `Bearer ${personData.token}`,
+            "auth-token": `${personData.token}`,
+          },
+        })
+          .then((response) => {
+            if (!response.ok) {
+              throw new Error("Network response was not ok");
+            }
+            return response.json();
+          })
+          .then((data) => {
+            Swal.fire({
+              title: "Package has been deleted",
+              showCancelButton: false,
+            });
+          })
+          .catch((error) => {
+            console.error("Error fetching data:", error);
+          });
+          window.location.reload();
+      }
+    });
   }
+  
   console.log(cardData, hall);
 
   useEffect(() => {
@@ -77,29 +90,30 @@ const PackageCard = (props) => {
           {/* {isOwner&&(<FaTrash onClick={deleteHall}  />)} */}
         </div>
 
-        <p>{cardData.package_description}</p>
+        <p className="custom-scrollbar">{cardData.package_description}</p>
         <p>Price: {cardData.price} </p>
         <div className="package-btn">
-          <Link
-            to={`/hallDetails/${cardData.hall_id}`}
-            onClick={handleHallDetailsClick}
+          <button
+          className="buttonMain details"
+           
+            onClick={()=>handleHallDetailsClick(hall)}
           >
             Details
-            {showHallDetails && (
+            {/* {showHallDetails && (
               <HallProfile
                 showHallDetails={showHallDetails}
                 cardData={cardData}
               />
-            )}
-          </Link>
-          {personData && (
-            <button className="Details-button" onClick={handleClick}>
-              Edit Package
+            )} */}
+          </button>
+          {personData.role==="admin" && (
+            <button className="buttonMain edit" onClick={handleClick}>
+              Edit 
             </button>
           )}
-          {personData && (
-            <button className="Details-button" onClick={deletePackage}>
-              Delete Package
+          {personData.role==="admin" && (
+            <button className="buttonMain delete" onClick={deletePackage}>
+              Delete 
             </button>
           )}
         </div>
